+++ tower/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
-None.
+- **builder**: Add `ServiceBuilder::layer_fn` to add a layer built from a function.
 
 # 0.4.5 (February 10, 2021)